requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
duckduckgo_search==6.2.11
fastapi==0.114.2
gitdb==4.0.11
GitPython==3.1.43
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
phi==0.6.7
phidata==2.4.34
primp==0.6.2
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.5
tomli==2.0.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
